% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4clinstats-package.R
\docType{package}
\name{r4clinstats-package}
\alias{r4clinstats}
\alias{r4clinstats-package}
\title{r4clinstats: Learn R for Clinical Statistics, Interactively}
\description{
A swirl-style, learn-by-doing course that teaches clinicians and clinical researchers to wrangle, tidy, visualise and analyse clinical data in R. Lessons are interactive 'learnr' tutorials graded by 'gradethis', built on the tidyverse and real-feeling synthetic clinical datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/apsmunro/r4clinstats}
  \item Report bugs at \url{https://github.com/apsmunro/r4clinstats/issues}
}

}
\author{
\strong{Maintainer}: A. P. S. Munro \email{apsmunro@gmail.com}

Authors:
\itemize{
  \item A. P. S. Munro \email{apsmunro@gmail.com}
}

}
\keyword{internal}
